Move PasswordEntry tests into entry_test.go
